New v3.2
Removed the walk forward feature. Since I don't see much difference at all in results without it, decided to remove it as this reduces the time it takes to generate horizon forecasts (by a hellofalot!) Still trying to get the prediction matplots to generate as intended, but this is an extra new feature I'm working on. The predictions are still being generated same as before, printed out in terminal or stored as csv correctly with better accuracy than v3.1

Signed-off-by: solo-coder-bro <kev90bourn@gmail.com>